# backend/src/main/resources/api.yaml
openapi: 3.0.2


info:
  title: Swagger Petstore - OpenAPI 3.0
  description: |-
    Something about your project.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0-SNAPSHOT
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io

paths:
  /store/health:
    get:
      summary: Check the status of the backend.
      operationId: sayHelloToWorld
      responses:
        200:
          description: "Literally returns 'Hello world'. Innovative, I know."
          content:
            text/plain:
              schema:
                type: string
  /store/item/{id}:
    get:
      summary: Get the information of an item.
      operationId: getItemById
      parameters:
        - name: id
          in: path 
          schema:
            type: integer
          required: true
      responses:
        200:
          description: "Returns the item info."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemInfo'
  /store/item:
    get:
      summary: "Retuns all items"
      operationId: getAllItems
      responses:
        200:
          description: All items in database
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllItems'

  /store/item/orderPrice:
    post:
      summary: Get the total price of the order.
      operationId: calcOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Order"
        description: configuration data to be appended at the current level
        required: true
      responses:
        200:
          description: "Returns the order total price."
          content:
            application/json:
              schema:
                type: number
  /store/login:
    post:
      summary: Authorize a login.
      operationId: authLogin
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginCredentials"
        description: configuration data to be appended at the current level
        required: true
      responses:
        200:
          description: "Authorized login."
        401:
          description: "Error with credentials."

components:
  schemas:
    ItemInfo:
      type: object
      properties:
        id:
          type: integer
          description: the ID of the item.
        name:
          type: string
          description: the name of the item.
        price:
          type: number
          description: the price of the item.
    Order:
      type: object
      properties:
        prices:
          type: array
          description: The item prices.
          items:
              type: number
    LoginCredentials:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    AllItems:
      type: array
      items:
          type: object
          properties:
            name:
              type: string
            price:
              type: number
            img:
              type: string
            id:
              type: integer